transMap pipeline.
Pipeline supports the basic beginnings of extracting genome and annotation files, then running chaining and transMap.

Decided on a simpler execution model, see issue #3.

Implemented add-ons for luigi to handle multiple inheritance and an interface with toil. This interface with toil is not as robust as I wish it was.