Added data munging to gff3 input.
See issue #5. This fix will make Ensembl gff3 parseable in the same way as Gencode. Some information is lost in columns we don't look at, like the description column. This could be fancier and only split on the transcript: and gene: markers. But this will work to getting what we want - a consistent input genePred for transMap, and a attributes TSV that can be dumped to sql.