Transcript alignment module.
Introduced the transcript alignment module. This module performs the task of generating a transcript level alignment for all transcripts, returning a PSL that will be parsed by the classification module. If AugustusCGP is ran, all possible transcripts (all transcripts with the biotype 'protein_coding' associated with the gene ID in the name2 field) will be aligned. This module lead to the introduction of two new concepts:

1. Instead of storing the attributes as a tsv, it is immediately imported into a sqlite database. A series of sqlite interface methods were added, and placed in tools.sqlInterface.py.
2. Convenience functions for loading/unloading flattened fastas were added to the Toil module in order to simplify code that loads the three components of a pyfasta flattened fast. This was applied to the existing Augustus code as well.
3. Imported the nameConversion.py code from the old pipeline.
4. Realized that the temporary file part of fileOps would not produce unique names if the files were not created before a subsequent call. Now random numbers are added, which should hopefully avoid a collision, but it is not guaranteed.