Generate hints DB, Toil bug fixes.
Added a tool to the pipeline that executes separately with its own Luigi workflow. This tool uses a config file to generate a RNAseq hints database.

In the process of designing this tool, I discovered that if I set --workDir on toil that my usage of getLocalTempDir() fails. I was informed that it is better for caching to use os.getcwd() anyways. As a result, I added a new function to fileOps to wrap this, and fixed all of my existing usages of temporary files in toil pipelines. I also discovered that we want to do 'file://' because we have absolute paths.

TODO: Right now, the hints pipeline says it fails despite working (and loading the hints twice). This is because we need to define a better way of measuring completion and fill out the output() function of BuildDb. Likely, this should involve checking first that the genome is in the species names table, then checking that the number of hints for that species matches the gff. An alternative would be to actually check each row for being in the database, but this seems slow. But the upside of this is that it would prevent edge cases.