Logging is far more sane now.
Nothing should be written to stdout/stderr unless something crashed. The downside to this is that we don't get the summary at the end, because for whatever reason Luigi puts the final report at INFO level.

There is still discrepancy in how config is done -- the level in logging.cfg is what is used for Luigi, while Toil must be passed a level directly through --logLevel.