Logging is even more sane.
Logging, as set by default, will now only log ERROR level logs for Toil/Luigi. Luigi logs will be sent to a file. INFO or higher level logs set by the user using the CAT logger will be send to stdout.

TODO: This does not entirely work. If I set the luigi-interface level to INFO, it goes both to stdout and to log.txt. I do not know why this is the case.