Parameter changes.
Refactored a combination of cat.py, luigiAddons.py and toilInterface.py to re-work the paradigm of how jobs get their parameters. Now, each job resolves their parameters separately. This was spurred by realizing that the problem with hashing was nested dicts. This was also what was leading to huge logs. By doing it this way, jobs now query the input HAL a lot of times to get the genomes, but this is an insignificant amount of computation.

This is also the first step towards actually using namespaces (without implementing a hashable namespace) instead of dicts in all argument passing, which will be great. It also better follows the Luigi paradigm - in theory, you could launch just the AlignTrancripts module for a specific genome with the right command line arguments.