Stop running MUSCLE on CGP transcripts.
This breaks the conceptual paradigm in place. There is no good way to compare this to the MUSCLE alignments produced for mRNAs. For consensus finding, if PRANK fails, we will have to discard CGP transcripts. This is probably OK since the main job of CGP is to find novel genes anyways.