Synteny score.
Implemented a synteny scoring metric for alignment classification. This process evaluates how many gene IDs are the same both up and down stream of a given transMap relative to the parent, up to 5 in each direction. As a result, a synteny score of 10 is best while 0 is worst. This can be used in consensus finding to try and assign fewer retroposed pseudogenes to their parents. Fixing this also found a bug in the interval library.