Changed hashing mechanism for Gff3ToAttrs.
We now directly hash the input gff3 file as the marker for table updates. This will prevent annoying rerunning of the Gff3ToAttrs module when the pipeline arguments signature changes. This approach works because the file is required to exist at the start, which means we can't use this for the other table updates, unfortunately.