Removing MUSCLE+FastTree.
MUSCLE proved to be a nightmare of edge cases. Using MAFFT helped. But in the end, when I did a density plot of the distances measured and the badness metric, it was almost the same thing. Instead, I have added a lognormal mixture model to help decide which alignment is to the correct paralog.

I also introduced a heuristic scoring metric for when this fails. This is a intuitive approach based on some examples I looked at and probably should have documented.

Right now, the pipeline is broken until I add in a new table definition for the results of this process. I also need to update the plots.