Updating homGeneMapping to perform exon mapping.
The homGeneMapping module now extracts wiggle hints from the hints database and uses those to populate the supplementary GTFs. The result of this is that we now have a measure of exon expression for consensus finding.

To support this, the ability to not provide a hints database has been removed. homGeneMapping is a integral part of the pipeline now because of the annotation support features. It may make sense to remove the separate step of hints database production.