Fixing split gene resolution.
Split gene resolution didn't really make sense, particularly in the context of how GFF3 files are validated. Now, they are evaluated regardless of the flag, and only removed if the flag is set. They also are evaluated on the same contig, and a boolean set if they are on the same one (and the alternate loci are filtered out).