Making parent gene assignment smarter.
Parent gene assignment previously was not its own task, which was a bad idea. It also was not very well designed. This improved design does a better job of filtering out false fusions. Consensus finding takes this into account and tries to now find actual fusions that are supported by RNA-seq.

TODO: There should be plots for the # of transcript types.