Reverting some AugustusPB changes.
After consulting with Stefanie and Mario, it was determined that simply turning off verbosity fixes the joingenes problem. AugustusPB can now be ran on chunks of genome, as before.

I also changed the classification scheme to use the RGB field to report the size of an indel for the indel tracks. This is useful for later parsing and figuring out compensatory frameshifts.

I also made the logging of parental gene assignment prettier.