Ancestral annotation, CGP training.
This update provides the ability to annotate ancestral genomes via the flag `--annotate-ancestors`. In addition, automated CGP training is now the default. Unless a pre-trained model is passed to `--cgp-param`, the pipeline will automatically train on the alignment. The amount of input to training can be modified by the `--cgp-train-num-exons` flag.

TODO NOTE: At the moment, if you run AugustusCGP with the `--annotate-ancestors` flag, the pipeline will fail. Stefanie is going to consider Benedict's workaround to making this happen.