More ancestral annotation updates.
Now that CGP training is integrated, it is possible to run CGP on a subset of genomes more easily, so this feature has been restored. Modifications to how target genomes are stored makes it possible to include them in the hub and plots. If `--target-genomes` and `--annotate-ancestors` is set, CAT will now intelligently find all common ancestors of the target genomes and annotate only those.