Fixes for ORF validation, GFF3.
To more properly support conversion to NCBI format, the classification database now records the validity of the start/stop codon for all transMap-derived transcripts (CGP/PB are guaranteed to have valid ORFs).

These data are now used to determine if the final GFF3 is given a start_codon and stop_codon feature, and are recorded in the gp_info files and the assembly hub. I also fixed the intervals for the GFF3 -- GFF3 CDS intervals *do* include the start/stop codons.

Also fixed a bug where multiple annotation sets in the config file will fail on hub building.