For issue #86.
The problem is once again that NCBI GFFs have biotype information not on the transcript level tag. This is resulting in my parser thinking that pseudogenes are protein coding. I hacked around this by inspecting the `pseudo=true` tag. I also fixed a problem with the genes with no transcript level features by inspecting to see if `gbkey=CDS`, and forcing those to protein coding. By my analysis of the A. thaliana GFF3, I am now seeing 27,636 protein coding genes in both my parser and the manual parsing of the GFF3.