Paralog filtering bug.
In the case where a given set of genes are present in multiple clusters, a transitive problem could occur when iterating and storing clusters to remove that you end up removing all instances of a gene family member. This bug fix instead works in the opposite direction, building up a list of transcripts to save instead of transcripts to remove.