Pull singularity image if using singularity
Singularity, unlike docker, keeps images in local files rather than
running a daemon that keeps track of available images and containers, so
this works a little bit differently than the same docker code. This
commit adds code for pulling the singularity image when the pipeline
begins and keeping track of where it is locally.